NEDI-6 fix : db_ui 부분 수정
